review updates:
- move implementation details into core
