+++ core/device_hooks/cuda_hooks.cpp
@@ -40,6 +40,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/version.hpp>
 
 
+#include "core/base/scoped_device_id.hpp"
 #include "core/matrix/csr_kernels.hpp"
